Include rendering before save per ajax request to plugin
